pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lacuna-wiki"
version = "2.0.0"
requires-python = ">=3.11"
dependencies = [
    "click>=8.0",
    "duckdb>=0.10.0",
    "rich>=13.0",
    "tomli-w>=1.0",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "watchdog>=3.0",
    "psutil>=5.9",
    "mcp>=1.0",
    "yt-dlp>=2024.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "respx>=0.20"]

[project.scripts]
lacuna = "lacuna_wiki.cli.main:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/lacuna_wiki"]
include = ["src/lacuna_wiki/skills/*.md"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
markers = [
    "integration: live network + yt-dlp (run with -m integration)",
]
